+++ packages/clerk-js/src/core/resources/Error.ts
@@ -3,6 +3,7 @@ export {
   ClerkAPIResponseError,
   EmailLinkError,
   EmailLinkErrorCode,
+  EmailLinkErrorCodeStatus,
   isClerkAPIResponseError,
   isClerkRuntimeError,
   isEmailLinkError,