+++ packages/nextjs/src/errors.ts
@@ -4,6 +4,7 @@ export {
   isKnownError,
   isMetamaskError,
   EmailLinkErrorCode,
+  EmailLinkErrorCodeStatus,
 } from './client-boundary/hooks';
 
 export { isClerkAPIResponseError } from '@clerk/clerk-react/errors';