+++ packages/shared/src/error.ts
@@ -210,11 +210,18 @@ export function isEmailLinkError(err: Error): err is EmailLinkError {
   return err.name === 'EmailLinkError';
 }
 
+/** @deprecated Please use `EmailLinkErrorCodeStatus` instead.*/
 export const EmailLinkErrorCode = {
   Expired: 'expired',
   Failed: 'failed',
   ClientMismatch: 'client_mismatch',
 };
+
+export const EmailLinkErrorCodeStatus = {
+  Expired: 'expired',
+  Failed: 'failed',
+  ClientMismatch: 'client_mismatch',
+} as const;
 
 const DefaultMessages = Object.freeze({
   InvalidProxyUrlErrorMessage: `The proxyUrl passed to Clerk is invalid. The expected value for proxyUrl is an absolute URL or a relative path with a leading '/'. (key={{url}})`,